requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==2.2.6
djangorestframework==3.10.3
pytz==2019.3              # via django
pyyaml==5.1.2
sqlparse==0.3.0           # via django
uritemplate==3.0.0
